Move c10:::ScalarType to a standalone header file
Summary:  Following the example of refactoring c10::DeviceType, move c10::ScalarType into a standalone header file.
