+++ c10/core/ScalarType.cpp
@@ -3,131 +3,6 @@
 #include <array>
 
 namespace c10 {
-
-namespace {
-
-constexpr auto u1 = ScalarType::Byte;
-constexpr auto i1 = ScalarType::Char;
-constexpr auto i2 = ScalarType::Short;
-constexpr auto i4 = ScalarType::Int;
-constexpr auto i8 = ScalarType::Long;
-constexpr auto f2 = ScalarType::Half;
-constexpr auto f4 = ScalarType::Float;
-constexpr auto f8 = ScalarType::Double;
-constexpr auto c2 = ScalarType::ComplexHalf;
-constexpr auto c4 = ScalarType::ComplexFloat;
-constexpr auto c8 = ScalarType::ComplexDouble;
-constexpr auto b1 = ScalarType::Bool;
-constexpr auto bf = ScalarType::BFloat16;
-constexpr auto ud = ScalarType::Undefined;
-
-constexpr auto index2dtype = array_of<
-    c10::ScalarType>(u1, i1, i2, i4, i8, f2, f4, f8, c2, c4, c8, b1, bf);
-
-constexpr std::array<int64_t, static_cast<size_t>(ScalarType::NumOptions)>
-calculate_dtype2index() {
-  std::array<int64_t, static_cast<size_t>(ScalarType::NumOptions)> inverse = {};
-  for (int64_t i = 0; i < static_cast<int64_t>(ScalarType::NumOptions); i++) {
-    inverse[i] = -1;
-  }
-  for (int64_t i = 0; i < static_cast<int64_t>(index2dtype.size()); i++) {
-    inverse[static_cast<int64_t>(index2dtype[i])] = i;
-  }
-  return inverse;
-}
-
-constexpr auto dtype2index = calculate_dtype2index();
-
-} // anonymous namespace
-
-ScalarType promoteTypes(ScalarType a, ScalarType b) {
-  // This is generated according to NumPy's promote_types
-  if (a == ud || b == ud) {
-    return ScalarType::Undefined;
-  }
-
-  // If the two types are equal, return that type
-  if (a == b) {
-    return a;
-  }
-
-  // Handle identically equal types
-  if (isQIntType(a) || isQIntType(b)) {
-    TORCH_CHECK(
-        false,
-        "promoteTypes with quantized numbers is not handled yet; figure out what the correct rules should be, offending types: ",
-        toString(a),
-        " ",
-        toString(b));
-  }
-
-  if (isBitsType(a) || isBitsType(b)) {
-    return ScalarType::Undefined;
-  }
-
-  if (isFloat8Type(a) || isFloat8Type(b)) {
-    TORCH_CHECK(
-        false,
-        "Promotion for Float8 Types is not supported, attempted to promote ",
-        toString(a),
-        " and ",
-        toString(b));
-  }
-
-  if (isBarebonesUnsignedType(a) || isBarebonesUnsignedType(b)) {
-    // There are two problems with promotion here:
-    //
-    // - Our promotion rule for uint8 is inconsistent with Numpy; Numpy
-    //   promotes to uint64, but since we never had uint64 for the longest
-    //   time, we promote to int64.  Changing this is BC-breaking
-    //
-    // - We must not promote uint64 to int64 because this will overflow.
-    //
-    // It'll be a bit of work to fix it, so we're punting on it for now.
-    // However, float promotion is fine, so we handle that.
-    if (isFloatingType(a)) {
-      return a;
-    }
-    if (isFloatingType(b)) {
-      return b;
-    }
-    TORCH_CHECK(
-        false,
-        "Promotion for uint16, uint32, uint64 types is not supported, attempted to promote ",
-        toString(a),
-        " and ",
-        toString(b));
-  }
-
-  auto ix_a = dtype2index[static_cast<int64_t>(a)];
-  TORCH_INTERNAL_ASSERT(ix_a != -1);
-  auto ix_b = dtype2index[static_cast<int64_t>(b)];
-  TORCH_INTERNAL_ASSERT(ix_b != -1);
-
-  // This table axes must be consistent with index2dtype
-  // clang-format off
-  static constexpr std::
-  array<std::array<ScalarType, index2dtype.size()>, index2dtype.size()>
-      _promoteTypesLookup = {{
-      /*        u1  i1  i2  i4  i8  f2  f4  f8  c2  c4  c8  b1  bf*/
-      /* u1 */ {u1, i2, i2, i4, i8, f2, f4, f8, c2, c4, c8, u1, bf},
-      /* i1 */ {i2, i1, i2, i4, i8, f2, f4, f8, c2, c4, c8, i1, bf},
-      /* i2 */ {i2, i2, i2, i4, i8, f2, f4, f8, c2, c4, c8, i2, bf},
-      /* i4 */ {i4, i4, i4, i4, i8, f2, f4, f8, c2, c4, c8, i4, bf},
-      /* i8 */ {i8, i8, i8, i8, i8, f2, f4, f8, c2, c4, c8, i8, bf},
-      /* f2 */ {f2, f2, f2, f2, f2, f2, f4, f8, c2, c4, c8, f2, f4},
-      /* f4 */ {f4, f4, f4, f4, f4, f4, f4, f8, c4, c4, c8, f4, f4},
-      /* f8 */ {f8, f8, f8, f8, f8, f8, f8, f8, c8, c8, c8, f8, f8},
-      /* c2 */ {c2, c2, c2, c2, c2, c2, c4, c8, c2, c4, c8, c2, c4},
-      /* c4 */ {c4, c4, c4, c4, c4, c4, c4, c8, c4, c4, c8, c4, c4},
-      /* c8 */ {c8, c8, c8, c8, c8, c8, c8, c8, c8, c8, c8, c8, c8},
-      /* b1 */ {u1, i1, i2, i4, i8, f2, f4, f8, c2, c4, c8, b1, bf},
-      /* bf */ {bf, bf, bf, bf, bf, f4, f4, f8, c4, c4, c8, bf, bf},
-  }};
-  // clang-format on
-  return _promoteTypesLookup[ix_a][ix_b];
-}
-
 std::pair<std::string, std::string> getDtypeNames(c10::ScalarType scalarType) {
   switch (scalarType) {
     case c10::ScalarType::UInt1:
